// src/lib/types/generated/AiConfigDiscoveryResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentConfig } from "./AgentConfig";
import type { DiscoverySource } from "./DiscoverySource";
import type { HookConfig } from "./HookConfig";
import type { InstructionConfig } from "./InstructionConfig";
import type { McpServerConfig } from "./McpServerConfig";
import type { MemoryConfig } from "./MemoryConfig";
import type { RuleConfig } from "./RuleConfig";
import type { SkillConfig } from "./SkillConfig";

export type AiConfigDiscoveryResult = { skills: Array<SkillConfig>, agents: Array<AgentConfig>, hooks: Array<HookConfig>, mcp_servers: Array<McpServerConfig>, memories: Array<MemoryConfig>, instructions: Array<InstructionConfig>, rules: Array<RuleConfig>, sources: Array<DiscoverySource>, };